# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: OGDEN Atlas API
  description: |
    Geospatial land intelligence API for the OGDEN Land Design Atlas.
    Manages projects, data layers, design features, PDF exports, public portals,
    real-time collaboration (WebSocket), and AI enrichment.
  version: 0.2.0
  contact:
    name: OGDEN Atlas
    url: https://atlas.ogden.ag
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List all projects for the authenticated user (owned + shared)
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProjectSummary'

    post:
      tags: [Projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
            example:
              name: Whispering Oaks Farm
              description: 100-acre regenerative land design
              country: US
              units: imperial
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectSummary'
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get project details (any project role)
      responses:
        '200':
          description: Full project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          description: Project not found

    patch:
      tags: [Projects]
      summary: Update project metadata (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectInput'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'

    delete:
      tags: [Projects]
      summary: Delete project permanently (owner only, cascades)
      responses:
        '204':
          description: Project deleted
  /projects/{id}/boundary:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Projects]
      summary: Set or replace parcel boundary (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [geojson]
              properties:
                geojson:
                  type: object
                  description: GeoJSON MultiPolygon or Polygon
      responses:
        '200':
          description: Boundary set; returns updated acreage and centroid
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          acreage:
                            type: number
                          hasParcelBoundary:
                            type: boolean
  /projects/{id}/assessment:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get site assessment with terrain analysis (any project role)
      responses:
        '200':
          description: Assessment data (null if pipeline not yet complete)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /projects/{id}/completeness:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get data completeness score (any project role)
      responses:
        '200':
          description: Completeness breakdown
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          score:
                            type: number
                            nullable: true
                          layers:
                            type: array
                            items:
                              type: object
  /projects/{id}/my-role:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Get the current user's role on this project
      responses:
        '200':
          description: Role string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          role:
                            $ref: '#/components/schemas/ProjectRole'

  # ── Members ────────────────────────────────────────────────────────────────
...
  /layers/project/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Layers]
      summary: List all data layers for a project (any project role)
      responses:
        '200':
          description: Layer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /layers/project/{projectId}/{layerType}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: layerType
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Layers]
      summary: Get a specific data layer (any project role)
      responses:
        '200':
          description: Layer data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          description: Layer not found
...
  /design-features/project/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Design Features]
      summary: List all design features for a project (any project role)
      responses:
        '200':
          description: Feature list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/DesignFeature'

    post:
      tags: [Design Features]
      summary: Create a design feature (owner or designer)
      description: Broadcasts a feature_created WebSocket event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignFeatureInput'
      responses:
        '201':
          description: Feature created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DesignFeature'
  /design-features/project/{projectId}/{featureType}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: featureType
        in: path
        required: true
        schema:
          type: string
          enum: [zone, structure, path, point, annotation]
    get:
      tags: [Design Features]
      summary: List features filtered by type (any project role)
      responses:
        '200':
          description: Filtered feature list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
...
  /spiritual/project/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Spiritual]
      summary: List spiritual zones for a project (any project role)
      responses:
        '200':
          description: Spiritual zone list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'

    post:
      tags: [Spiritual]
      summary: Create a spiritual zone (owner or designer)
      description: |
        Automatically computes Qibla bearing for prayer_space and qibla_axis
        zone types using the project centroid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [zoneType, geometry]
              properties:
                zoneType:
                  type: string
                  enum: [prayer_space, qibla_axis, contemplation_garden, wudu_station, cemetery, sacred_grove]
                geometry:
                  type: object
                  description: GeoJSON geometry
                name:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Spiritual zone created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
...
  /ws/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: token
        in: query
        required: true
        schema:
          type: string
        description: |
          JWT token (same token from /auth/login or /auth/register).
          Passed as query parameter because browser WebSocket API cannot send
          custom Authorization headers on upgrade.
    get:
      tags: [WebSocket]
      summary: Open a real-time collaboration WebSocket for a project
      description: |
        This endpoint upgrades to a WebSocket connection. Authentication is
        performed using the JWT passed as the `token` query parameter. Only
        project owners and members may connect.

        **On connection:**
        - Server broadcasts `presence_join` to other connected users
        - Server sends `presence_sync` to the new connection with list of
          currently connected users

        **Client → Server messages:**
        | type | payload |
        |---|---|
        | `presence_heartbeat` | (none) |
        | `typing_start` | `{ action: string }` |
        | `typing_stop` | (none) |

        **Server → Client events:**
        | type | description |
        |---|---|
        | `feature_created` | Design feature added |
        | `feature_updated` | Design feature modified |
        | `feature_deleted` | Design feature removed |
        | `comment_added` | New comment |
        | `comment_resolved` | Comment marked resolved |
        | `comment_deleted` | Comment deleted |
        | `layer_complete` | Data pipeline layer finished |
        | `export_ready` | PDF export completed |
        | `presence_join` | User connected |
        | `presence_leave` | User disconnected |
        | `presence_sync` | Full presence list (sent on connect only) |
        | `typing_start` | User started editing |
        | `typing_stop` | User stopped editing |

        **Close codes:**
        - `4001` Missing or invalid token
        - `4003` Not a project member
        - `4004` Project not found
        - `4009` Replaced by new connection from same user
        - `4500` Server error during auth
      responses:
        '101':
          description: WebSocket upgrade successful
        '401':
          description: Missing or invalid token (upgrade rejected)
        '403':
          description: Not a member of this project (upgrade rejected)
        '404':
          description: Project not found (upgrade rejected)

  # ── Health ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/register or /auth/login (expires in 7 days)
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          type: object
          properties:
            total:
              type: integer
        error:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
...
    ProjectRole:
      type: string
      enum: [owner, designer, reviewer, viewer]
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        projectType:
          type: string
          nullable: true
        country:
          type: string
        provinceState:
          type: string
          nullable: true
        conservationAuthId:
          type: string
          nullable: true
        units:
          type: string
        address:
          type: string
          nullable: true
        parcelId:
          type: string
          nullable: true
        acreage:
          type: number
          nullable: true
        dataCompletenessScore:
          type: number
          nullable: true
        hasParcelBoundary:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProjectInput:
      type: object
      required: [name, country, units]
      properties:
        name:
          type: string
        description:
          type: string
        projectType:
          type: string
        country:
          type: string
        provinceState:
          type: string
        units:
          type: string
          enum: [imperial, metric]
    UpdateProjectInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        projectType:
          type: string
        address:
          type: string
        parcelId:
          type: string
        ownerNotes:
          type: string
        zoningNotes:
          type: string
        accessNotes:
          type: string
        waterRightsNotes:
          type: string
...
    DesignFeature:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        featureType:
          type: string
          enum: [zone, structure, path, point, annotation]
        subtype:
          type: string
          nullable: true
        geometry:
          type: object
          description: GeoJSON geometry
        label:
          type: string
          nullable: true
        properties:
          type: object
        phaseTag:
          type: string
          nullable: true
        style:
          type: object
          nullable: true
        sortOrder:
          type: integer
        createdBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDesignFeatureInput:
      type: object
      required: [featureType, geometry, properties, sortOrder]
      properties:
        featureType:
          type: string
          enum: [zone, structure, path, point, annotation]
        subtype:
          type: string
        geometry:
          type: object
        label:
          type: string
        properties:
          type: object
        phaseTag:
          type: string
        style:
          type: object
        sortOrder:
          type: integer
...